+++ src/huc/function.h
@@ -7,13 +7,13 @@
 #define _FUNCTION_H
 
 void newfunc (void);
-void getarg (int t);
+void getarg (long t);
 void callfunction (char *ptr);
-void arg_stack(int arg);
+void arg_stack(long arg);
 void arg_push_ins(INS *ptr);
-void arg_flush(int arg, int adj);
-void arg_to_fptr(struct fastcall *fast, int i, int arg, int adj);
-void arg_to_dword(struct fastcall *fast, int i, int arg, int adj);
+void arg_flush(long arg, long adj);
+void arg_to_fptr(struct fastcall *fast, long i, long arg, long adj);
+void arg_to_dword(struct fastcall *fast, long i, long arg, long adj);
 
 #endif
